delete old Pill code
