--- packages/lab/src/pill/PillNext.css
@@ -1,183 +1,0 @@
-/* Styles applied to the root element */
-.saltPill {
-  --pill-background: var(--saltPill-background, var(--salt-taggable-background));
-  --pill-background-active: var(--saltPill-background-active, var(--salt-taggable-background-active));
-  --pill-background-disabled: var(--saltPill-background-disabled, var(--salt-taggable-background-disabled));
-  --pill-background-hover: var(--saltPill-background-hover, var(--salt-taggable-background-hover));
-  --pill-fontSize: var(--saltPill-fontSize, var(--salt-text-fontSize));
-  --pill-height: calc(var(--salt-size-base) - var(--salt-size-unit));
-  --pill-icon-color: var(--saltPill-icon-color, var(--salt-taggable-foreground));
-  --pill-icon-color-active: var(--saltPill-icon-color-active, var(--salt-taggable-foreground-active));
-  --pill-icon-color-hover: var(--saltPill-icon-color-hover, var(--salt-taggable-foreground-hover));
-  --pill-icon-color-disabled: var(--saltPill-icon-color-disabled, var(--salt-taggable-foreground-disabled));
-  --pill-text-color: var(--saltPill-text-color, var(--salt-taggable-foreground));
-  --pill-text-color-active: var(--saltPill-text-color-active, var(--salt-taggable-foreground-active));
-  --pill-text-color-hover: var(--saltPill-text-color-hover, var(--salt-taggable-foreground-hover));
-  --pill-text-color-disabled: var(--saltPill-text-color-disabled, var(--salt-taggable-foreground-disabled));
-}
-
-.saltPill {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  background: var(--pill-background);
-  border-radius: var(--saltPill-borderRadius, 0);
-  height: var(--saltPill-height, var(--pill-height));
-  max-width: var(--saltPill-maxWidth, 160px);
-  min-width: var(--saltPill-minWidth, 40px);
-  min-height: var(--saltPill-minHeight, var(--salt-text-minHeight));
-  outline: var(--saltPill-outline, 0);
-  position: relative;
-  gap: var(--saltPill-gap, var(--salt-spacing-50));
-  padding-inline: var(--salt-spacing-50);
-  color: var(--saltPill-text-color, var(--pill-text-color));
-  font-size: var(--pill-fontSize);
-  font-weight: var(--saltPill-fontWeight, var(--salt-text-fontWeight));
-  letter-spacing: var(--saltPill-letterSpacing, 0);
-  line-height: var(--saltPill-lineHeight, var(--salt-text-label-lineHeight));
-}
-
-.saltPill-label {
-  min-width: 0;
-  overflow: hidden;
-  text-align: center;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.saltPill-closable {
-  padding-right: 0;
-}
-
-.saltPill .saltIcon {
-  --saltIcon-color: var(--pill-icon-color);
-}
-
-.saltPill-closeButton .saltIcon {
-  display: block;
-}
-
-/* Style applied to Pill if `clickable={true}`  */
-.saltPill-clickable {
-  cursor: var(--salt-selectable-cursor-hover);
-  user-select: none;
-}
-
-.saltPill-clickable:hover,
-.saltPill-closeButton:hover {
-  --saltIcon-color: var(--pill-icon-color);
-  --saltIcon-color-hover: var(--pill-icon-color-hover);
-
-  background: var(--pill-background-hover);
-}
-
-.saltPill-clickable[aria-disabled]:hover,
-.saltPill-nestedHover:hover {
-  --saltIcon-color: var(--pill-icon-color);
-  --saltIcon-color-hover: var(--pill-icon-color);
-
-  background: var(--pill-background);
-}
-
-.saltPill-clickable.saltPill-active,
-.saltPill-clickable:active {
-  background: var(--pill-background-active);
-  color: var(--pill-text-color-active);
-}
-
-.saltPill-clickable.saltPill-active .saltIcon,
-.saltPill-clickable:active .saltIcon {
-  --saltIcon-color: var(--pill-text-color-active);
-  --saltIcon-color-hover: var(--pill-text-color-active);
-}
-
-.saltPill-nestedHover.saltPill-active .saltIcon,
-.saltPill-nestedHover:active .saltIcon {
-  --saltIcon-color: var(--pill-icon-color);
-  --saltIcon-color-hover: var(--pill-icon-color);
-}
-
-.saltPill-clickable .saltPill-closeButton.saltPill-active .saltIcon,
-.saltPill-clickable .saltPill-closeButton:active .saltIcon {
-  --saltIcon-color: var(--pill-text-color-active);
-  --saltIcon-color-hover: var(--pill-text-color-active);
-}
-
-.saltPill-clickable[aria-disabled].saltPill-active,
-.saltPill-clickable[aria-disabled]:active,
-.saltPill-clickable.saltPill-nestedHover.saltPill-active,
-.saltPill-clickable.saltPill-nestedHover:active {
-  background: var(--pill-background);
-  color: var(--pill-text-color);
-  --saltIcon-color: var(--pill-text-color);
-}
-
-/* Style applied to Pill on focus */
-.saltPill:focus-visible {
-  outline-color: var(--saltPill-focused-outlineColor, var(--salt-focused-outlineColor));
-  outline-style: var(--saltPill-focused-outlineStyle, var(--salt-focused-outlineStyle));
-  outline-width: var(--salt-focused-outlineWidth);
-  outline-offset: var(--salt-focused-outlineOffset);
-}
-
-/* Style applied to Pill when disabled */
-.saltPill[aria-disabled],
-.saltPill[aria-disabled]:hover,
-.saltPill[aria-disabled]:active,
-.saltPill[aria-disabled] .saltPill-closeButton,
-.saltPill[aria-disabled] .saltPill-closeButton:hover,
-.saltPill[aria-disabled] .saltPill-closeButton:active {
-  background: var(--pill-background-disabled);
-  cursor: var(--salt-selectable-cursor-disabled);
-  opacity: var(--salt-palette-opacity-foreground);
-
-  --saltIcon-color: var(--pill-icon-color-disabled);
-  --saltIcon-color-hover: var(--pill-icon-color-disabled);
-}
-
-.saltPill[aria-disabled] .saltPill-closeButton,
-.saltPill[aria-disabled] .saltPill-closeButton:hover,
-.saltPill[aria-disabled] .saltPill-closeButton:active {
-  background: transparent;
-}
-
-/* Style applied to Pill delete button */
-.saltPill-closeButton {
-  --saltIcon-size-multiplier: 1;
-  flex-shrink: 0;
-  height: var(--pill-height);
-  width: var(--pill-height);
-  min-height: var(--saltPill-minHeight, var(--salt-text-minHeight));
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.saltPill-closeButton:hover {
-  cursor: var(--salt-selectable-cursor-hover);
-  background: var(--salt-actionable-secondary-background-hover);
-  color: var(--salt-actionable-secondary-foreground-hover);
-}
-
-.saltPill-closeButton.saltPill-active,
-.saltPill-closeButton:active {
-  background: var(--pill-background-active);
-  color: var(--pill-text-color-active);
-  --saltIcon-color: var(--pill-text-color-active);
-}
-
-/* Style applied to Pill delete button icon on active state  */
-.saltPill-closeButton:active .saltPill-deleteIcon {
-  --saltIcon-color: var(--saltPill-closeButton-color-active, var(--pill-icon-color-active));
-}
-
-/* Style applied to Pill delete button icon if `disabled={true}` on active state  */
-.saltPill-disabled:active .saltPill-deleteIcon {
-  --saltIcon-color: var(--saltPill-closeButton-color-disabled, var(--pill-icon-color));
-}
-
-/* Style applied to Pill icon on active state */
-.saltPill-clickable:not([aria-disabled]) .saltPill-icon,
-.saltPill-clickable.saltPill-active:not([aria-disabled]) .saltPill-icon {
-  fill: var(--pill-icon-color-active);
-}
