--- packages/lab/src/pill/constants.ts
@@ -1,1 +1,0 @@
-export const pillBaseName = "saltPill";
