+++ packages/lab/src/pill/index.ts
@@ -1,3 +1,1 @@
-export * from "./PillNext";
-// export * from "./ClosablePill";
-// export * from "./SelectablePill";
+export * from "./Pill";
