--- packages/lab/src/pill/internal/PillCheckbox.css
@@ -1,14 +1,0 @@
-/* Styles applied to the Pill checkbox */
-.saltPill-checkbox {
-  height: var(--pill-checkbox-size);
-  margin-left: 1px;
-  padding-right: 0;
-  width: var(--pill-checkbox-size);
-
-  --saltCheckbox-icon-stroke: var(--salt-selectable-borderColor);
-  --saltCheckbox-icon-stroke-hover: var(--salt-selectable-borderColor);
-  --saltCheckbox-icon-fill: none;
-  --saltCheckbox-icon-fill-checked: none;
-  --saltCheckbox-icon-borderWidth-checked: var(--salt-size-border);
-  --saltCheckbox-icon-tick-color: var(--salt-taggable-foreground);
-}
